refactor: move model parsing inside reference service
Less separation of related code.
